requirements.txt
# === CORE FRAMEWORK (PyTorch + MPS) ===
torch==2.6.0
torchaudio==2.6.0 # Re-added as FFmpeg issues are resolved
#  --extra-index-url https://download.pytorch.org/whl/nightly/cpu # Commented out to prioritize MPS support
torchcodec

# === TTS MODELS ===
piper-tts==1.3.0
TTS==0.22.0  # Coqui TTS (zero-shot cloning with XTTS_v2)


# === STT: Faster-Whisper (PyTorch + CTranslate2 + MPS) ===
faster-whisper==1.2.0
ctranslate2>=4.0.0

# === MT: SeamlessM4T v2 or NLLB (Hugging Face Transformers) ===
transformers==4.46.3
sentencepiece==0.2.1
accelerate==0.33.0

# === AUDIO & UTILS ===
librosa==0.11.0
soundfile==0.13.1
numpy==1.26.4
scipy==1.13.1
numba==0.60.0
av==15.1.0
pydub==0.25.1
ffmpeg-python==0.2.0 # Added for FFmpeg integration
jiwer==4.0.0
sacrebleu==2.5.1
pandas==1.5.3 # Adjusted for compatibility with TTS==0.22.0
scikit-learn==1.7.2
nltk==3.9.2

# === COMMON DEPENDENCIES ===
tokenizers==0.20.3
tqdm==4.67.1
huggingface-hub==0.36.0
safetensors==0.5.3
pyyaml==6.0.3
regex==2025.10.23
requests==2.32.5
packaging==25.0
jinja2==3.1.6
markupsafe==3.0.3
typing-extensions==4.15.0
filelock==3.19.1
fsspec==2025.10.0

# === WEB / API ===
fastapi==0.120.3
uvicorn[standard]==0.38.0
websockets==12.0
httpx==0.28.1 # Added for backend testing
python-multipart # Required for form data handling in FastAPI

# === ONNX Runtime (for Piper) ===
onnxruntime==1.19.2

# === VOICE ACTIVITY DETECTION (VAD) ===
webrtcvad==2.0.10

# === Optional: Pre-commit, etc. ===
pre-commit==4.3.0
argon2-cffi==23.1.0
pytest
pytest-asyncio
sacremoses

# === DATABASE ===
SQLAlchemy==2.0.30
alembic==1.13.1
python-jose[cryptography]==3.3.0

# === UI Testing ===
playwright==1.45.0
